requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.2.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.6
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
cryptography==41.0.5
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.58.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.2.0
frozenlist==1.4.0
graphviz==0.20.1
idna==3.4
inflection==0.5.1
isodate==0.6.1
msal==1.26.0
msal-extensions==1.1.0
multidict==6.0.4
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
portalocker==2.8.2
psycopg2-binary==2.9.9
pycparser==2.21
pydot==1.4.2
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
railway==0.0.4
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
yarl==1.9.2